updated parameter of positive checker to double from int to accept double and ints

+++ Capstone/Utilities/UserInt.cs
@@ -3,7 +3,7 @@ namespace Capstone.Utilities
 {
     public class UserInt
     {
-        public static bool IsPositiveNumber(int num)
+        public static bool IsPositiveNumber(double num)
         {
             if (num >= 0) return true;
             else return false;
